Made selection a normal ListProperty again
Copied the batch_delete() method from OpObservableDict to
ObservableDict, so that the selection property can be the normal type of
ListProperty, dispatching change events normally. This makes the API for
binding to selection a normal affair. The detailed change info that came
from using OpObservableList was not used in the examples. This can be
used if there is a need for it in the future.